+++ Backend/src/main/java/com/eminimal/backend/utils/FileService.java
@@ -1,4 +1,4 @@
-package com.eminimal.backend.services.impl;
+package com.eminimal.backend.utils;
 
 import com.dropbox.core.DbxException;
 import com.dropbox.core.DbxRequestConfig;
@@ -17,13 +17,17 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 @Service
-public class FileServiceImpl implements com.eminimal.backend.services.interfaces.FileService {
-    private static final String ACCESS_TOKEN = "sl.BRxn-jTGQFTbp789B9BZs0E1EToN4wMPLpjct6LLkLAqFn0TE97GbJ03Z0ASg6QnuabKLOwx_SqF_DIBZ91Ge5n4MecghhnsgqrSEXdvowVWMlNBtlVTa2zSs9M7yVn86ySk0e7pHhk";
-    private static final Logger logger = LoggerFactory.getLogger(FileServiceImpl.class);
-    
-    @Override
+public class FileService {
+    private static final String ACCESS_TOKEN = "sl.BShgWz0Kc4H9eKFUghqND8TeWSGCZXfaNTQdTw0HtzkK1Fh1J7KH9k6V_Ti5X0MCdpZb4hX5TvJw9UsdbuHJQiKGScqQ6n_XNCjYOoYPs280Br7I84q6F6cvn3G6rUZp0-5WiVLFTi8";
+    private static final String pathDropbox = "/upload/images/";
+
+
+    private static final Logger logger = LoggerFactory.getLogger(FileService.class);
+
     public String upload(File file) throws Exception {
 //		Dropbox
         DbxRequestConfig config = DbxRequestConfig.newBuilder("App/eMinimal").build();
@@ -49,8 +53,11 @@ public class FileServiceImpl implements com.eminimal.backend.services.interfaces
 
         // Upload file to Dropbox
         try (InputStream in = new FileInputStream(file.getPath())) {
+//            new SimpleDateFormat("ddMMyy-hhmmss.SSS-file.txt")
+            String timeStamp = new SimpleDateFormat("yyyy-MM-dd_HH:mm:ss").format(new Date());
 
-            String path = "/upload/images/" + file.getName();
+            logger.error("File name: " + timeStamp);
+            String path = "/upload/images/" + timeStamp  + "_" + file.getName();
             FileMetadata metadata = client.files().uploadBuilder(path).uploadAndFinish(in);
 
             SharedLinkMetadata sharedLinkMetadata = client.sharing().createSharedLinkWithSettings(path);
@@ -61,4 +68,5 @@ public class FileServiceImpl implements com.eminimal.backend.services.interfaces
             throw new Exception(e.getMessage());
         }
     }
+
 }
